+++ src/SegmentFinder.h
@@ -1,6 +1,7 @@
 #ifndef SEGMENT_FINDER_H_INCLUDED
 #define SEGMENT_FINDER_H_INCLUDED
 
+#include <cstdint>
 #include "Segment.h"
 
 #include <queue>
@@ -59,7 +60,7 @@ private:
 #pragma pack(pop)
 
 	Node    *m_root;
-	size_t   m_num_objs;
+	uint64_t   m_num_objs;
 	unsigned m_max_depth;
 	unsigned m_max_node_objs;
